Add updating to PropertyQueries

+++ init.sql
@@ -2,8 +2,8 @@
 -- hasTenants is 0 if there are no tenants, 1 if there are
 CREATE TABLE properties (
     id VARCHAR(10) NOT NULL,
-    balance INT,
-    price INT,
+    balance FLOAT,
+    price FLOAT,
     moveIn DATE,
     hasTenants INT,
     description VARCHAR(500),
